move objects definition to seprate file
